+++ src/pinsController.hpp
@@ -12,7 +12,7 @@ struct Pin {
     uint32_t number;
     PinType type;
     PinDirection direction;
-    float value;
+    uint32_t value;
 };
 
 class PinsController {
@@ -24,12 +24,12 @@ public:
     void setPin(Pin pin);
     void setPinType(uint32_t pinNum, PinType type);
     void setPinDirection(uint32_t pinNum, PinDirection direction);
-    void setPinValue(uint32_t pinNum, float value);
+    void setPinValue(uint32_t pinNum, uint32_t value);
 
     Pin getPin(uint32_t pinNum);
     PinType getPinType(uint32_t pinNum);
     PinDirection getPinDirection(uint32_t pinNum);
-    float getPinValue(uint32_t pinNum);
+    uint32_t getPinValue(uint32_t pinNum);
 
 private:
     std::vector<Pin> mPins;
